moved summary card to widget: It looks like independent block
