+++ .codemap/float-area-b05f199e-963b-477b-a4e9-baf34e603fec.json
@@ -36,7 +36,52 @@
                         },
                         "contributionsState": {
                             "editor.contrib.folding": {
-                                "lineCount": 10,
+                                "lineCount": 17,
+                                "foldedImports": false
+                            },
+                            "editor.contrib.wordHighlighter": false
+                        }
+                    }
+                },
+                "width": 600,
+                "height": 450,
+                "spacing": 0
+            },
+            {
+                "widget": {
+                    "constructionOptions": {
+                        "factoryId": "code-editor-opener",
+                        "options": {
+                            "kind": "navigatable",
+                            "uri": "src/pages/new/ui/New.tsx",
+                            "floatAreaId": "b05f199e-963b-477b-a4e9-baf34e603fec"
+                        }
+                    },
+                    "innerWidgetState": {
+                        "cursorState": [
+                            {
+                                "inSelectionMode": false,
+                                "selectionStart": {
+                                    "lineNumber": 1,
+                                    "column": 1
+                                },
+                                "position": {
+                                    "lineNumber": 1,
+                                    "column": 1
+                                }
+                            }
+                        ],
+                        "viewState": {
+                            "scrollLeft": 0,
+                            "firstPosition": {
+                                "lineNumber": 1,
+                                "column": 1
+                            },
+                            "firstPositionDeltaTop": 0
+                        },
+                        "contributionsState": {
+                            "editor.contrib.folding": {
+                                "lineCount": 117,
                                 "foldedImports": false
                             },
                             "editor.contrib.wordHighlighter": false
