+++ src/pages/app/ui/OverviewTab.tsx
@@ -1,5 +1,5 @@
 import { Navbar, Link } from "konsta/react";
-import { SummaryCard } from "@/features/summary-card";
+import { SummaryCard } from "@/widgets/summary-card";
 import { MdExitToApp } from "react-icons/md";
 import { Preferences } from "@/entities/preferences";
